Fixed localization for input elements using plain objects for translation value.
Before this change, if you tried to using a plain object as the localization
value for input elements (e.g. trying to localize the title), then the value
would be set to "[Object object]".

After this change, you can set the translation value to an object with +title+
and +value+ properties that will be used to set the title and placeholder of the
input element.