Using raw object as data source
New code has been added so that raw javascript objects can now be used as
a source for the translations.

If localize receives an object as a first parameter, items with
'data-localize' tags will be updated with the text contained in the
matching object properties. Said object will be converted to a valid json
object beforehand, so no problems shall arise if the given object include
functions as a value for any of its keys.

In the case localize method receives any parameter that is not an object,
the default behaviour where the translations are retrieved from an
external file will be applied.

This should solve the following issue:
https://github.com/coderifous/jquery-localize/issues/62