Support for custom callback when using raw objects
The method that replace the localized tags by the texts defined inside of
a raw object did not consider any of the possible options you could
receive as a second parameter in a 'localize' call.

The 'callback' option is now supported. The 'data' object will be
converted to a valid json object before calling the default callback, so
if the user sets any of the data properties to a function, the elements
for which the 'data-localize' attribute references any of those invalid property
values will not be updated.

The 'localize_test.coffee' file do now include new methods to test the new functionality.

The 'localize_from_object.html' has been also updated
and seems to be working as expected in the following browsers:

 - Google Chrome (59.0.3071.115)
 - Mozilla Firefox (45.0, 47.0.1)
 - Microsoft Edge (40.15063.0.0)