Added new example for localize with objects
I've added a new example to illustrate on how to call localize by passing
the response of a webservice as an argument to 'jquery-localize'.

Both of the examples created for this new feature have been moved to a new
'objectDS' folder, as to keep them separated from the default examples.