Removed function properties in sample object
Since we want the examples to be fully functional and show how the library
is meant to be used, I've removed the properties for which I've assigned a
function as their value.

Still, the sanitizer callback will be kept and so will the test to
guarantee that there will be no problem if the user passes an object that
include a function among their keys.